patch 8.2.2826: compiler warnings for int to size_t conversion
Problem:    Compiler warnings for int to size_t conversion. (Randall W.
            Morris)
Solution:   Add type casts.

+++ src/map.c
@@ -1552,7 +1552,7 @@ check_abbr(
 			escaped = vim_strsave_escape_csi(tb + j);
 			if (escaped != NULL)
 			{
-			    newlen = STRLEN(escaped);
+			    newlen = (int)STRLEN(escaped);
 			    mch_memmove(tb + j, escaped, newlen);
 			    j += newlen;
 			    vim_free(escaped);
